Made it so every time a Pokemon is shown, its name is split up into letters that the user must arrange to make up the Pokemon's name. There are always 16 letters to choose from, made up from the Pokemon's name and random letters to fill up the rest. There are frames behind the letters at the bottom and frames at the user input area where the letters go when pressed. If someone clicks on a letter in the user input area, it is taken back to its original location. The app keeps track of which spots are open at the top so that whenever a button at the bottom is clicked, it takes up the leftmost open spot.
When "Check" is clicked, the app checks if it is the correct answer. If it is correct, then "Correct!" is printed, and the next time the user clicks the button it will give them a new Pokemon to guess. If it is incorrect, "Incorrect!" is printed.

Also made it so the new Pokemon information only loads after the Pokemon image is ready (won't be necessary in the future, as the image won't appear at the start, unless we decide to add a pixelated silhouette or something of the like).

As my previous app, Slanguage (https://github.com/dylantelson/Slanguage) used a similar user input method, much of the user input code was reused and modified to fit the needs of this app.