Added account creation and login using Firebase. This implementation borrows some code from my previous project, Slanguage. I learned how to do this in Firebase this for Slanguage using the following article: https://medium.com/@ashikabala01/how-to-build-login-and-sign-up-functionality-for-your-ios-app-using-firebase-within-15-mins-df4731faf2f7
Added Startup, Login, and Signup ViewControllers, each with its own purpose with account handling. Also modified the constraints a lot- it is still a little messed up, and is based simply on distance from top/left/etc. rather than taking the height and width into account, and as thus still is not working correctly in screens other than the iPhone 11 Pro Max. The code for the user input buttons etc. also has to be modified so it appears based on the height of the screen rather than a fixed Y which is too large for most devices.