Added anonymous signup. Now the user can play the game without signing up just by clicking "Anonymous Login" and is given its firebase user ID as their name. They can still use their account as if it was real, can make it to the leaderboards, and  have their login saved even if they close and reopen the app, but if they ever logout they can never retrieve their account.
Fixed leaderboard. Now the users with the top 5 scores have their name and score listed in the leaderboard, and if the currently logged in user is in it their name is blue to distinguish it. Using this method, the "Name" form should be replaced with a username that must be unique when signing up.

Added the logo to the leaderboard and changed the background color of the leaderboard view.

Slightly changed the Y value of the first row of letters to better fit the design and not overlap with the visual line.